Calculate Unlocked Balance

+++ pkg/wallet/wallet.go
@@ -176,34 +176,39 @@ func (w *Wallet) NewBidTx(fee int64, lockTime uint64, amount ristretto.Scalar) (
 	return tx, nil
 }
 
-func (w *Wallet) CheckWireBlock(blk block.Block) (uint64, uint64, error) {
-	spentCount, err := w.CheckWireBlockSpent(blk)
+// CheckWireBlock
+// update if true, walletDB state is updated accordingly
+func (w *Wallet) CheckWireBlock(blk block.Block, update bool) (uint64, uint64, error) {
+	spentCount, err := w.CheckWireBlockSpent(blk, update)
 	if err != nil {
 		return 0, 0, err
 	}
 
-	receivedCount, err := w.CheckWireBlockReceived(blk)
+	receivedCount, _, err := w.CheckWireBlockReceived(blk, update)
 	if err != nil {
 		return 0, 0, err
 	}
 
-	err = w.UpdateWalletHeight(blk.Header.Height + 1)
-	if err != nil {
-		return 0, 0, err
+	if update {
+		err = w.UpdateWalletHeight(blk.Header.Height + 1)
+		if err != nil {
+			return 0, 0, err
+		}
 	}
+
 	return spentCount, receivedCount, nil
 }
 
 // CheckWireBlockSpent checks if the block has any outputs spent by this wallet
 // Returns the number of txs that the sender spent funds in
-func (w *Wallet) CheckWireBlockSpent(blk block.Block) (uint64, error) {
+func (w *Wallet) CheckWireBlockSpent(blk block.Block, update bool) (uint64, error) {
 
 	var totalSpentCount uint64
 
 	txInCheckers := NewTxInChecker(blk)
 
 	for _, txchecker := range txInCheckers {
-		spentCount, err := w.scanInputs(txchecker)
+		spentCount, err := w.scanInputs(txchecker, update)
 		if err != nil {
 			return spentCount, err
 		}
@@ -213,7 +218,7 @@ func (w *Wallet) CheckWireBlockSpent(blk block.Block) (uint64, error) {
 	return totalSpentCount, nil
 }
 
-func (w *Wallet) scanInputs(txChecker TxInChecker) (uint64, error) {
+func (w *Wallet) scanInputs(txChecker TxInChecker, update bool) (uint64, error) {
 
 	var didSpendFunds uint64
 
@@ -226,45 +231,54 @@ func (w *Wallet) scanInputs(txChecker TxInChecker) (uint64, error) {
 			return didSpendFunds, err
 		}
 
-		err = w.db.RemoveInput(pubKey, keyImage)
-		if err != nil {
-			return didSpendFunds, err
+		didSpendFunds++
+
+		if update {
+			err = w.db.RemoveInput(pubKey, keyImage)
+			if err != nil {
+				return didSpendFunds, err
+			}
 		}
+
 	}
 	return didSpendFunds, nil
 }
 
 // CheckWireBlockReceived checks if the wire block has transactions for this wallet
 // Returns the number of tx's that the reciever recieved funds in
-func (w *Wallet) CheckWireBlockReceived(blk block.Block) (uint64, error) {
+// update if true, walletDB state is updated accordingly
+func (w *Wallet) CheckWireBlockReceived(blk block.Block, update bool) (uint64, uint64, error) {
 
 	var totalReceivedCount uint64
 
 	txCheckers := NewTxOutChecker(blk)
+	var totalBalance uint64
 	for _, txchecker := range txCheckers {
-		receivedCount, err := w.scanOutputs(txchecker)
+		receivedCount, balance, err := w.scanOutputs(txchecker, update)
 		if err != nil {
-			return receivedCount, err
+			return totalReceivedCount, totalBalance, err
 		}
 		totalReceivedCount += receivedCount
+		totalBalance += balance
 	}
 
-	return totalReceivedCount, nil
+	return totalReceivedCount, totalBalance, nil
 }
 
 // scans the outputs of one transaction
-func (w *Wallet) scanOutputs(txchecker TxOutChecker) (uint64, error) {
+func (w *Wallet) scanOutputs(txchecker TxOutChecker, update bool) (uint64, uint64, error) {
 
 	privView, err := w.keyPair.PrivateView()
 	if err != nil {
-		return 0, err
+		return 0, 0, err
 	}
 	privSpend, err := w.keyPair.PrivateSpend()
 	if err != nil {
-		return 0, err
+		return 0, 0, err
 	}
 
 	var didReceiveFunds uint64
+	var totalAmount uint64
 
 	for i, output := range txchecker.Outputs {
 		privKey, ok := w.keyPair.DidReceiveTx(txchecker.R, output.PubKey, uint32(i))
@@ -283,23 +297,28 @@ func (w *Wallet) scanOutputs(txchecker TxOutChecker) (uint64, error) {
 			mask = transactions.DecryptMask(output.EncryptedMask, txchecker.R, uint32(i), *privView)
 		}
 
-		err := w.db.PutInput(privSpend.Bytes(), output.PubKey.P, amount, mask, *privKey)
-		if err != nil {
-			return didReceiveFunds, err
+		totalAmount += amount.BigInt().Uint64()
+
+		if update {
+			err := w.db.PutInput(privSpend.Bytes(), output.PubKey.P, amount, mask, *privKey)
+			if err != nil {
+				return didReceiveFunds, totalAmount, err
+			}
+
+			// cache the keyImage, so we can quickly check whether our input was spent
+			var pubKey ristretto.Point
+			pubKey.ScalarMultBase(privKey)
+			keyImage := mlsag.CalculateKeyImage(*privKey, pubKey)
+
+			err = w.db.Put(keyImage.Bytes(), output.PubKey.P.Bytes())
+			if err != nil {
+				return didReceiveFunds, totalAmount, err
+			}
 		}
 
-		// cache the keyImage, so we can quickly check whether our input was spent
-		var pubKey ristretto.Point
-		pubKey.ScalarMultBase(privKey)
-		keyImage := mlsag.CalculateKeyImage(*privKey, pubKey)
-
-		err = w.db.Put(keyImage.Bytes(), output.PubKey.P.Bytes())
-		if err != nil {
-			return didReceiveFunds, err
-		}
 	}
 
-	return didReceiveFunds, nil
+	return didReceiveFunds, totalAmount, nil
 }
 
 // AddInputs adds up the total outputs and fee then fetches inputs to consolidate this
